Initialized database 'medimage2report.db'
